# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ReplayDash API",
    "description": "Session replay and user analytics API for capturing and analyzing user interactions",
    "version": "1.0",
    "contact": {
      "name": "ReplayDash Support",
      "url": "https://replaydash.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    },
    {
      "url": "https://api.replaydash.com",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/api/v1/events": {
      "post": {
        "tags": ["events"],
        "summary": "Ingest session events",
        "description": "Captures and stores user interaction events for session replay. Includes DOM snapshots, clicks, scrolls, and other user interactions.",
        "operationId": "ingestEvents",
        "security": [
          {
            "api-key": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestEventsDto"
              },
              "examples": {
                "basic": {
                  "summary": "Basic event ingestion",
                  "value": {
                    "sessionId": "sess_abc123def456",
                    "userId": "user_789xyz",
                    "userEmail": "user@example.com",
                    "userAgent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36",
                    "url": "https://example.com/dashboard",
                    "timestamp": 1708777200000,
                    "events": [
                      {
                        "type": "rrweb/full-snapshot",
                        "timestamp": 1708777200000,
                        "data": {
                          "node": {
                            "type": 0,
                            "childNodes": []
                          },
                          "initialOffset": {
                            "top": 0,
                            "left": 0
                          }
                        }
                      },
                      {
                        "type": "rrweb/mouse-move",
                        "timestamp": 1708777200100,
                        "data": {
                          "x": 150,
                          "y": 200
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Events successfully ingested",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "API key for authentication. Get your API key from the ReplayDash dashboard."
      }
    },
    "schemas": {
      "IngestEventsDto": {
        "type": "object",
        "required": ["sessionId", "events", "timestamp"],
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Unique session identifier",
            "example": "sess_abc123def456"
          },
          "userId": {
            "type": "string",
            "description": "User identifier (if authenticated)",
            "example": "user_789xyz"
          },
          "userEmail": {
            "type": "string",
            "description": "User email address (if available)",
            "example": "user@example.com"
          },
          "userAgent": {
            "type": "string",
            "description": "Browser user agent string",
            "example": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36"
          },
          "url": {
            "type": "string",
            "description": "URL of the page where events were captured",
            "example": "https://example.com/dashboard"
          },
          "events": {
            "type": "array",
            "description": "Array of captured events",
            "items": {
              "$ref": "#/components/schemas/EventDto"
            }
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp in milliseconds for the batch",
            "example": 1708777200000
          }
        }
      },
      "EventDto": {
        "type": "object",
        "required": ["type", "timestamp", "data"],
        "properties": {
          "type": {
            "type": "string",
            "description": "Event type (e.g., rrweb/full-snapshot, rrweb/mouse-move)",
            "example": "rrweb/full-snapshot"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp in milliseconds when the event occurred",
            "example": 1708777200000
          },
          "data": {
            "type": "object",
            "description": "Event data payload (structure varies by event type)",
            "additionalProperties": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Invalid request data"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      }
    }
  }
}